AI Team Support/requirements.txt
openai==0.28
notion-client
pandas
python-dateutil
gradio
scikit-learn
numpy